requirements.txt
django
djangorestframework
beautifulsoup4
requests
pandas
scikit-learn
lxml
numpy